respa_admin: Move login view to separate module
